refactor(cli): move schedule request builders to schedule_request.go (PRD #1009 M2)
